implemented reading Zip64 pack file;

+++ Code/CryEngine/CryCommon/CrySystem/File/ICryPak.h
@@ -393,7 +393,7 @@ struct ICryPak
 	virtual char const* const GetLocalizationFolder() const = 0;
 
 	//! Only returns useful results on a dedicated server at present - and only if the pak is already opened.
-	virtual void GetCachedPakCDROffsetSize(const char* szName, uint32& offset, uint32& size) = 0;
+	virtual void GetCachedPakCDROffsetSize(const char* szName, uint64& offset, uint64& size) = 0;
 
 	//! \return Array of PackInfo structures inside OpenPacks structure. You MUST call FreeOpenPackInfo.
 	virtual ICryPak::PakInfo* GetPakInfo() = 0;
