+++ Code/CryEngine/CrySystem/ZipDirCacheRW.h
@@ -153,7 +153,7 @@ protected:
 
 	// offset to the start of CDR in the file,even if there's no CDR there currently
 	// when a new file is added, it can start from here, but this value will need to be updated then
-	uint32            m_lCDROffset;
+	uint64            m_lCDROffset;
 
 	CSimpleStringPool m_tempStringPool;
